requirements.txt
django>=5.0.0,<5.2.1
djangorestframework
pyyaml
requests
django-cors-headers